llm_rag_tutorial/chatbot_api/pyproject.toml
[project]
name = "chatbot-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "black>=25.1.0",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-huggingface>=0.3.0",
    "langchain-neo4j>=0.4.0",
    "langchain-ollama>=0.3.3",
    "langchainhub>=0.1.21",
    "neo4j>=5.28.1",
    "numpy>=2.3.1",
    "openai>=1.91.0",
    "opentelemetry-api>=1.34.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=4.1.0",
    "uvicorn>=0.34.3",
]
